# swagger.yaml
...
openapi: 3.0.3
info:
  title: CCP Disposizione Bonifico
  description: |+
    Interfaccia per la disposizione di un bonifico cross border (fuori dall'area SEPA).

    ## Macchina a stati bonifico

    La seguente macchina a stati descrive i possibili stati in cui si trova un bonifico.

    Lo stato inserted e to be confirmed sono previsti per il flusso di inserimento da frontend di backoffice.

    [![](https://mermaid.ink/img/pako:eNplUk1PAyEU_CsvHE17sN_hYFKrBxN7qomJrofnwrYk7KNhoUab_ndxd8HScmKGGd5A5shKIyTjrHHo5IPCrcV6eBgVVBCEtbnl8ESNtE6KnhlxeDFwL2FlqFK2TgdjDo_WGpuISVQuvdsZq37SyZTD6-oZyDjYY9MkfsZhhVRKrRM157CR5Hq04NeXzWA4vIP3m48YucV_OZfkFFQWvQAkAQfUSqBThvKpvWGcGYw9118lHV8MHXV3hCenb2lnfpW6Mza5cHou_PSuFV5N6ZPNcm-EkxxOczi_SNirF_EnOzjPzSH_Gsmjjuna58c2sAGrZeCUCI05_nEFcztZy4LxsBWyQq9dwQo6BSl6ZzbfVDLurJcD5vfiv2OMV6ibwEqhnLHrroVtGQfMGr_dJcUe6c2Y6Dj9AtBuzIo?type=png)](https://mermaid.live/edit#pako:eNplUk1PAyEU_CsvHE17sN_hYFKrBxN7qomJrofnwrYk7KNhoUab_ndxd8HScmKGGd5A5shKIyTjrHHo5IPCrcV6eBgVVBCEtbnl8ESNtE6KnhlxeDFwL2FlqFK2TgdjDo_WGpuISVQuvdsZq37SyZTD6-oZyDjYY9MkfsZhhVRKrRM157CR5Hq04NeXzWA4vIP3m48YucV_OZfkFFQWvQAkAQfUSqBThvKpvWGcGYw9118lHV8MHXV3hCenb2lnfpW6Mza5cHou_PSuFV5N6ZPNcm-EkxxOczi_SNirF_EnOzjPzSH_Gsmjjuna58c2sAGrZeCUCI05_nEFcztZy4LxsBWyQq9dwQo6BSl6ZzbfVDLurJcD5vfiv2OMV6ibwEqhnLHrroVtGQfMGr_dJcUe6c2Y6Dj9AtBuzIo)

    ### Esempi

...


...

  version: 1.0.0
...
paths:
  /bonifico/banca:
    post:
      tags:
      - Bonifici
      summary: Crea un nuovo bonifico banca a banca
      description: |
        Il body della richiesta contiene tutti i campi necessari per inserire un nuovo bonifico nel sistema banca a banca.

        Il canale dal quale arriva la richiesta invece è riconosciuto dalle informazioni di autenticazione fornite.

        Il canale di ingresso permette di inserire un bonifico in uno stato diverso, configurabile per banca. è possibile inoltre modificare lo stato iniziale in cui il bonifico viene inserito, lo stato iniziale non puo' essere uno stato più avanzato rispetto allo stato iniziale di default del canale.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InserisciBonificoExtraSepaBancaRichiesta"
      responses:
        "201":
          description: Bonifico creato con successo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BonificoExtraSepaRisposta"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /bonifico/cliente:
    post:
      tags:
      - Bonifici
      summary: Crea un nuovo bonifico account a account
      description: |
        Il body della richiesta contiene tutti i campi necessari per inserire un nuovo bonifico nel sistema account to account.

        Il canale dal quale arriva la richiesta invece è riconosciuto dalle informazioni di autenticazione fornite.

        Il canale di ingresso permette di inserire un bonifico in uno stato diverso, configurabile per banca. è possibile inoltre modificare lo stato iniziale in cui il bonifico viene inserito, lo stato iniziale non puo' essere uno stato più avanzato rispetto allo stato iniziale di default del canale.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InserisciBonificoExtraSepaClienteRichiesta"
      responses:
        "201":
          description: Bonifico creato con successo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BonificoExtraSepaRisposta"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
  /bonifico/{id}:
    get:
      tags:
      - Bonifici
      summary: Ottiene lo stato di un bonifico
      description: |
        Ottiene lo stato attuale di un bonifico.
      parameters:
      - name: id
        in: path
        description: L'id del bonifico
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BonificoExtraSepaRisposta"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
    patch:
      tags:
      - Bonifici
      summary: Modifica un bonifico
      description: |
        Il body della richiesta contiene i campi da aggiornare.

        La transizione tra il nuovo e precedente stato deve essere permesso dalla macchina a stati dei bonifici.
      parameters:
      - name: id
        in: path
        description: L'id del bonifico da modificare
        required: true
        schema:
          $ref: "#/components/schemas/UUID"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RichiestaModificaBonificoStato"
      responses:
        "200":
          description: Bonifico modificato con successo.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BonificoExtraSepaRisposta"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme: []
...
components:
  schemas:
...
    AltriIntermediari:
      type: object
      properties:
        riferimentiCorrispondenteMittente:
          $ref: "#/components/schemas/Rapporto"
        bancaCorrispondenteMittente:
          $ref: "#/components/schemas/Intermediario"
        bancaCorrispondenteRicevente:
          $ref: "#/components/schemas/Intermediario"
        istitutoTerzoDiRimborso:
          $ref: "#/components/schemas/Intermediario"
        bancaIstruttrice1:
          $ref: "#/components/schemas/Intermediario"
        bancaIstruttrice2:
          $ref: "#/components/schemas/Intermediario"
        bancaIstruttrice3:
          $ref: "#/components/schemas/Intermediario"
        bancaIntermediaria1:
          $ref: "#/components/schemas/Intermediario"
        bancaIntermediaria2:
          $ref: "#/components/schemas/Intermediario"
        bancaIntermediaria3:
          $ref: "#/components/schemas/Intermediario"
    Attore:
      description: Informazioni su un attore
      type: object
      properties:
        intestazione:
          type: string
        indirizzo:
          $ref: "#/components/schemas/Indirizzo"
        paeseDiResidenza:
          description: Paese di residenza
          type: string
        organizzazione:
          $ref: "#/components/schemas/InfoOrganizzazione"
        privato:
          $ref: "#/components/schemas/InfoPrivato"
        rapporto:
          $ref: "#/components/schemas/Rapporto"
...
    BonificoExtraSepaRisposta:
      title: Dati di un bonifico nel sistema
      description: Contiene tutti i campi relativi a un bonifico nel sistema
      required:
      - id
      - tid
      - infoStato
      type: object
      properties:
        id:
          description: Identificativo univoco del bonifico
          type: string
          allOf:
          - $ref: "#/components/schemas/UUID"
        tid:
          description: Identificativo della transazione
          type: string
        bonificoCliente:
          $ref: "#/components/schemas/DatiBonificoExtraSepaClienteRisposta"
        bonificoBanca:
          $ref: "#/components/schemas/DatiBonificoExtraSepaBancaRisposta"
        infoStato:
          description: Informazioni sullo stato del bonifico
          type: object
          allOf:
          - $ref: "#/components/schemas/InfoStato"
...
    CausaleBanca:
      description: |
        Causale di una transazione banca a banca.
      required:
      - causaleDescrittiva
      type: object
      properties:
        causaleDescrittiva:
          description: Rappresenta l'intera causale.
          pattern: \S
          type: string
...
    ContoBancaDiCopertura:
      title: Conto banca di copertura
      description: |
        Dati definenti il conto della banca coprente il bonifico.
      type: object
      properties:
        rapportoBanca:
          description: IBAN del conto della banca di copertura
          type: string
          example: BY86AKBB10100000002966000000
        divisa:
          description: Divisa della banca di copertura
          type: string
          example: euro
        bic:
          description: BIC della banca di copertura
          type: string
          example: PRACIT31XXX
        intestazione:
          description: Intestazione conto della banca di copertura
          type: string
    CreditTransferStatus:
      enum:
      - TO_BE_MANAGED
      - INSERITO
      - DA_CONFERMARE
      - CONFERMATO
      - ELIMINATO
      - WCL_NON_PASSATA
      - WCL_NEGATA
      - DA_AUTORIZZARE
      - AUTORIZZATO
      - INVIATO
      - REGOLATO
      - IN_ERRORE
      - NON_AUTORIZZATO_RIMBORSO_PROGRAMMATO
      - NON_AUTORIZZATO
      - RIFIUTATO_RIMBORSO_PROGRAMMATO
      - RIFIUTATO
      - STORNO_RICHIESTO
      - STORNO_INVIATO
      - STORNO_ACCETTATO
      - STORNO_RIFIUTATO
      - STORNO_RIMBORSO_PROGRAMMATO
      - RIMBORSATO
      type: string
    DatiBonificoExtraSepaBancaRisposta:
      description: Dati di un bonifico ExtraSepa tra banche già inserito nel sistema
      required:
      - sottoTipologiaBonifico
      - ordinante
      - bancaOrdinante
      - beneficiario
      - bancaDestinataria
      - bancaDelBeneficiario
      - altriIntermediari
      - sistemaDiRegolamento
      - dettagliBonifico
      - dettagliCausale
      - riferimentiAggiuntivi
      type: object
      properties:
        sottoTipologiaBonifico:
          $ref: "#/components/schemas/SottoTipologiaBonificoRisposta"
        ordinante:
          $ref: "#/components/schemas/Intermediario"
        bancaOrdinante:
          $ref: "#/components/schemas/Intermediario"
        beneficiario:
          $ref: "#/components/schemas/Intermediario"
        bancaDestinataria:
          $ref: "#/components/schemas/Intermediario"
        bancaDelBeneficiario:
          $ref: "#/components/schemas/Intermediario"
        altriIntermediari:
          $ref: "#/components/schemas/AltriIntermediari"
        sistemaDiRegolamento:
          $ref: "#/components/schemas/InfoSistemaDiRegolamento"
        dettagliBonifico:
          $ref: "#/components/schemas/DettagliBonificoExtraSepaBanca"
        dettagliCausale:
          $ref: "#/components/schemas/DettagliCausaleBanca"
        riferimentiAggiuntivi:
          $ref: "#/components/schemas/RiferimentiAggiuntivi"
        user:
          type: string
    DatiBonificoExtraSepaClienteRisposta:
      description: Dati di un bonifico ExtraSepa tra account già inserito nel sistema
      required:
      - sottoTipologiaBonifico
      - ordinante
      - soggettoIstruttore
      - debitoreEffettivo
      - bancaOrdinante
      - beneficiario
      - beneficiarioEffettivo
      - bancaDestinataria
      - bancaDelBeneficiario
      - altriIntermediari
      - sistemaDiRegolamento
      - dettagliBonifico
      - dettagliCausale
      - riferimentiAggiuntivi
      type: object
      properties:
        sottoTipologiaBonifico:
          $ref: "#/components/schemas/SottoTipologiaBonificoRisposta"
        ordinante:
          $ref: "#/components/schemas/InfoOrdinante"
        soggettoIstruttore:
          $ref: "#/components/schemas/Attore"
        debitoreEffettivo:
          $ref: "#/components/schemas/Attore"
        bancaOrdinante:
          $ref: "#/components/schemas/Intermediario"
        beneficiario:
          $ref: "#/components/schemas/Attore"
        beneficiarioEffettivo:
          $ref: "#/components/schemas/Attore"
        bancaDestinataria:
          $ref: "#/components/schemas/Intermediario"
        bancaDelBeneficiario:
          $ref: "#/components/schemas/Intermediario"
        altriIntermediari:
          $ref: "#/components/schemas/AltriIntermediari"
        sistemaDiRegolamento:
          $ref: "#/components/schemas/InfoSistemaDiRegolamento"
        dettagliBonifico:
          $ref: "#/components/schemas/DettagliBonificoExtraSepaCliente"
        dettagliCausale:
          $ref: "#/components/schemas/DettagliCausaleCliente"
        riferimentiAggiuntivi:
          $ref: "#/components/schemas/RiferimentiAggiuntivi"
        user:
          type: string
    DettagliBonificoExtraSepaBanca:
      required:
      - dettagliImporto
      - dettagliDate
      type: object
      properties:
        dettagliImporto:
          $ref: "#/components/schemas/DettagliImporto"
        dettagliDate:
          $ref: "#/components/schemas/DettagliDate"
        contoBancaDiCopertura:
          $ref: "#/components/schemas/ContoBancaDiCopertura"
        dettagliCommissioni:
          $ref: "#/components/schemas/DettagliCommissioni"
        dettagliNotifica:
          $ref: "#/components/schemas/DettagliNotifica"
    DettagliBonificoExtraSepaCliente:
      required:
      - dettagliImporto
      - dettagliDate
      type: object
      properties:
        dettagliImporto:
          $ref: "#/components/schemas/DettagliImporto"
        dettagliDate:
          $ref: "#/components/schemas/DettagliDate"
        contoBancaDiCopertura:
          $ref: "#/components/schemas/ContoBancaDiCopertura"
        dettagliCommissioni:
          $ref: "#/components/schemas/DettagliCommissioni1"
    DettagliCausaleBanca:
      title: Dettagli causale
      description: |
        Informazioni dettagliate sulla causale di un bonifico.
      type: object
      properties:
        codiceCausaleTransazione:
          title: Codice causale
          type: string
        causale:
          type: array
          items:
            $ref: "#/components/schemas/CausaleBanca"
    DettagliCausaleCliente:
      title: Dettagli causale
      description: |
        Informazioni dettagliate sulla causale di un bonifico.
      type: object
      properties:
        codiceCausaleTransazione:
          title: Codice causale
          type: string
        causale:
          type: array
          items:
            $ref: "#/components/schemas/CausaleCliente"
    DettagliCommissioni:
      type: object
      properties:
        commissioniBanca:
          $ref: "#/components/schemas/CommissioniBanca"
...
    DettagliDate:
      title: Dettagli date
      description: Informazioni sulle tempistiche del bonifico
      required:
      - dataCreazione
      - dataEsecuzione
      - dataValutaOrdinante
      - dataRegolamentoBancaBeneficiario
      type: object
      properties:
        dataCreazione:
          $ref: "#/components/schemas/LocalDate"
        dataEsecuzione:
          $ref: "#/components/schemas/LocalDate"
        dataValutaOrdinante:
          $ref: "#/components/schemas/LocalDate"
        dataRegolamentoBancaBeneficiario:
          $ref: "#/components/schemas/LocalDate"
    DettagliImporto:
      title: Dettagli importo
      description: Informazioni sull'importo del bonifico
      required:
      - importo
      - divisa
      - cambio
      type: object
      properties:
        importo:
          type: number
        divisa:
          type: string
        cambio:
          type: number
    DettagliNotifica:
      type: object
      properties:
        notifica:
          type: string
...
    Indirizzo:
      description: Un indirizzo fisico
      type: object
      properties:
        indirizzo:
          type: string
        citta:
          type: string
        cap:
          type: string
        paese:
          maxLength: 2
          type: string
        divisione:
          type: string
        sottoDivisione:
          type: string
        numeroCivico:
          type: string
        edificio:
          type: string
        piano:
          type: string
        cassettaPostale:
          type: string
        stanza:
          type: string
        localita:
          type: string
        distretto:
          type: string
        provincia:
          type: string
        lineaIndirizzo:
          type: string
...
    InfoOrdinante:
      type: object
      properties:
        infoAttore:
          $ref: "#/components/schemas/Attore"
        presentatore:
          type: array
          items:
            $ref: "#/components/schemas/InfoNDG"
        titolareEffettivo:
          type: array
          items:
            $ref: "#/components/schemas/InfoNDG"
...
    InfoSistemaDiRegolamento:
      required:
      - sistemaDiRegolamento
      type: object
      properties:
        sistemaDiRegolamento:
          $ref: "#/components/schemas/SistemaDiRegolamento"
        priorita:
          $ref: "#/components/schemas/Priorita"
        da:
          $ref: "#/components/schemas/Instant"
        a:
          $ref: "#/components/schemas/Instant"
        orarioAccredito:
          $ref: "#/components/schemas/Instant"
        scadenzaUltima:
          $ref: "#/components/schemas/Instant"
        stp:
          type: boolean
    InfoStato:
      required:
      - stato
      type: object
      properties:
        stato:
          description: Stato del bonifico
          type: string
          allOf:
          - $ref: "#/components/schemas/CreditTransferStatus"
...
    InserisciBonificoExtraSepaBancaRichiesta:
      description: |
        Dati necessari per inserire un bonifico banca a banca nel sistema.
      required:
      - sottoTipologiaBonifico
      - ordinante
      - bancaOrdinante
      - beneficiario
      - bancaDestinataria
      - bancaDelBeneficiario
      - altriIntermediari
      - sistemaDiRegolamento
      - dettagliBonifico
      - dettagliCausale
      - riferimentiAggiuntivi
      type: object
      properties:
        sottoTipologiaBonifico:
          $ref: "#/components/schemas/SottoTipologiaBonificoRichiesta"
        ordinante:
          $ref: "#/components/schemas/Intermediario"
        bancaOrdinante:
          $ref: "#/components/schemas/Intermediario"
        beneficiario:
          $ref: "#/components/schemas/Intermediario"
        bancaDestinataria:
          $ref: "#/components/schemas/Intermediario"
        bancaDelBeneficiario:
          $ref: "#/components/schemas/Intermediario"
        altriIntermediari:
          $ref: "#/components/schemas/AltriIntermediari"
        sistemaDiRegolamento:
          $ref: "#/components/schemas/InfoSistemaDiRegolamento"
        dettagliBonifico:
          $ref: "#/components/schemas/DettagliBonificoExtraSepaBanca"
        dettagliCausale:
          $ref: "#/components/schemas/DettagliCausaleBanca"
        riferimentiAggiuntivi:
          $ref: "#/components/schemas/RiferimentiAggiuntivi"
        user:
          type: string
    InserisciBonificoExtraSepaClienteRichiesta:
      description: |
        Dati necessari per inserire un bonifico account to account nel sistema.
      required:
      - sottoTipologiaBonifico
      - ordinante
      - soggettoIstruttore
      - debitoreEffettivo
      - bancaOrdinante
      - beneficiario
      - beneficiarioEffettivo
      - bancaDestinataria
      - bancaDelBeneficiario
      - altriIntermediari
      - sistemaDiRegolamento
      - dettagliBonifico
      - dettagliCausale
      - riferimentiAggiuntivi
      type: object
      properties:
        sottoTipologiaBonifico:
          $ref: "#/components/schemas/SottoTipologiaBonificoRichiesta"
        ordinante:
          $ref: "#/components/schemas/InfoOrdinante"
        soggettoIstruttore:
          $ref: "#/components/schemas/Attore"
        debitoreEffettivo:
          $ref: "#/components/schemas/Attore"
        bancaOrdinante:
          $ref: "#/components/schemas/Intermediario"
        beneficiario:
          $ref: "#/components/schemas/Attore"
        beneficiarioEffettivo:
          $ref: "#/components/schemas/Attore"
        bancaDestinataria:
          $ref: "#/components/schemas/Intermediario"
        bancaDelBeneficiario:
          $ref: "#/components/schemas/Intermediario"
        altriIntermediari:
          $ref: "#/components/schemas/AltriIntermediari"
        sistemaDiRegolamento:
          $ref: "#/components/schemas/InfoSistemaDiRegolamento"
        dettagliBonifico:
          $ref: "#/components/schemas/DettagliBonificoExtraSepaCliente"
        dettagliCausale:
          $ref: "#/components/schemas/DettagliCausaleCliente"
        riferimentiAggiuntivi:
          $ref: "#/components/schemas/RiferimentiAggiuntivi"
        user:
          type: string
    Instant:
      format: date-time
      type: string
      example: 2022-03-10T16:15:50Z
    Intermediario:
      type: object
      properties:
        bic:
          type: string
        intestazione:
          type: string
        codiceLEI:
          type: string
        codiceSistemaClearing:
          type: string
        identificativoClearing:
          type: string
        rapporto:
          $ref: "#/components/schemas/Rapporto"
        indirizzo:
          $ref: "#/components/schemas/Indirizzo"
...
    Priorita:
      enum:
      - URGENTE
      - ALTA
      - NORMALE
      type: string
...
    Rapporto:
      description: Rapporto tra la banca e un attore. Usualmente corrisponde a un
        conto aperto.
      required:
      - tipoRapporto
      type: object
      properties:
        numeroRapporto:
          description: Identificativo del rapporto
          type: string
          oneOf:
          - $ref: "#/components/schemas/NumeroRapportoIban"
          - $ref: "#/components/schemas/NumeroRapportoAltroIdentificativo"
        divisa:
          description: Divisa del conto
          type: string
        codiceTipoConto:
          description: Codice tipo conto
          type: string
        dettaglioTipoConto:
          description: Descrizione tipo conto
          type: string
        intestazioneConto:
          description: Intestazione del conto
          type: string
        codiceTipoAlias:
          description: Codice tipo identificativo alias
          type: string
        descrizioneAlias:
          description: Descrizione tipo identificativo alias
          type: string
        dettaglioIdentificativoAlias:
          description: Dettaglio tipo identificativo alias
          type: string
        codiceIdentificativoConto:
          description: Codice identificativo conto
          type: string
        descrizioneIdentificativoConto:
          description: Descrizione identificativo conto
          type: string
        emittente:
          description: Emittente del conto
          type: string
        codiceFiliale:
          description: Codice filiale della banca
          type: string
        denominazioneFiliale:
          description: Denominazione filiale della banca
          type: string
        tipoRapporto:
          $ref: "#/components/schemas/TipoRapporto"
        ndg:
          type: string
...
    RichiestaModificaBonificoStato:
      type: object
      properties:
        nuovoStato:
          $ref: "#/components/schemas/CreditTransferStatus"
    RiferimentiAggiuntivi:
      type: object
      properties:
        altroIdentificativoPagamento:
          type: string
        riferimentiAggiuntivi1:
          $ref: "#/components/schemas/DettaglioRiferimentiAggiuntivi"
        riferimentiAggiuntivi2:
          $ref: "#/components/schemas/DettaglioRiferimentiAggiuntivi"
        istruzioniBancaRicevente1:
          type: string
        istruzioniBancaRicevente2:
          type: string
        istruzioniBancaRicevente3:
          type: string
        istruzioniBancaRicevente4:
          type: string
        istruzioniBancaRicevente5:
          type: string
        istruzioniBancaRicevente6:
          type: string
        istruzioneBancaDelBeneficiario1:
          $ref: "#/components/schemas/CodiceEIstruzione"
        istruzioneBancaDelBeneficiario2:
          $ref: "#/components/schemas/CodiceEIstruzione"
        prioritaTransazione:
          $ref: "#/components/schemas/PrioritaTransazione"
        codiceLivelloDiServizio:
          type: string
        dettaglioLivelloDiServizio:
          type: string
        classificazionePagamento:
          type: string
        dettaglioClassificazionePagamento:
          type: string
        codiceTransazione:
          type: string
        dettaglioIdentificativoTransazione:
          type: string
        codiceTipoServizio:
          type: string
        dettaglioTipoServizio:
          type: string
        valoreCambioIstruito:
          type: number
        regulatoryReporting:
          type: array
          items:
            $ref: "#/components/schemas/InformazioniRegulatoryReporting"
    SistemaDiRegolamento:
      enum:
      - TARGET
      - NO_TARGET
      type: string
    SottoTipologiaBonificoRichiesta:
      title: Sottotipologia del bonifico
      description: |
        Sotto-tipologia per il bonifico “Area Extra Sepa”.

        Le sotto-tipologie sono configurabili per banca ma, al massimo, sono le seguenti:
        - Pacs.008
        - Pacs.008 con Pacs.009 COV
        - Pacs.008 con Pacs.009 COV e MT999
        - Pacs.008 con MT999
      required:
      - tipo
      type: object
      properties:
        tipo:
          pattern: \S
          type: string
          example: Pacs.008
...
    UUID:
      format: uuid
      pattern: "[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}"
      type: string
...